// src/schema/events/tests/editEvent.generated.ts
/* eslint-disable */
/* @ts-nocheck */
/* prettier-ignore */
/* This file is automatically generated using `npm run graphql:types` */
import type * as Types from '../../../generated/types';

import type { JsonObject } from "type-fest";
import gql from 'graphql-tag';
export type EditEventMutationVariables = Types.Exact<{
  input: Types.EventEditInput;
}>;


export type EditEventMutation = { __typename?: 'Mutation', editEvent: { __typename?: 'Event', id: string, name: string, description: string | null, visibility: Types.EventVisibility, status: Types.EventStatus, startDateTime: string, endDateTime: string | null } };


export const EditEvent = gql`
    mutation EditEvent($input: EventEditInput!) {
  editEvent(input: $input) {
    id
    name
    description
    visibility
    status
    startDateTime
    endDateTime
  }
}
    `;